requirements.txt
streamlit
pandas
plotly
openpyxl
xlrd==2.0.1
openai
pyyaml
msoffcrypto-tool
pdfplumber
